rename and add validation feedback on contact
