+++ frontend/package.json
@@ -44,6 +44,7 @@
     "vue-cli-plugin-apollo": "^0.21.3",
     "vue-jest": "^3.0.5",
     "vue-router": "^3.1.5",
+    "vuelidate": "^0.7.5",
     "vuex": "^3.1.2"
   },
   "devDependencies": {
